requirements.txt
numpy
opencv-python
rknn-toolkit2
onnx